Remove the old iterator implementation.
